+++ packages/vsce/src/webviews/resource-inspector-panel/ResourceInspector.tsx
@@ -84,7 +84,13 @@ const ResourceInspector = () => {
                 // @ts-ignore
                 resourceInfo ? (resourceInfo?.resource.status || resourceInfo?.resource.enablestatus) : "..."}
             </div>
-            {resourceInfo && <Contextmenu resourceActions={resourceActions} refreshIconPath={resourceInfo?.refreshIconPath} />}
+            {resourceInfo && (
+              <Contextmenu
+                resourceActions={resourceActions}
+                refreshIconPath={resourceInfo?.refreshIconPath}
+                resourceHumanReadableName={resourceInfo?.humanReadableNameSingular}
+              />
+            )}
           </th>
         </thead>
         <tbody className="padding-left-10 padding-top-20">