+++ packages/vsce/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 
 ## Recent Changes
 
-- Remove show attributes from tree menu [#422](https://github.com/zowe/cics-for-zowe-client/issues/422)
+- Replaced the Show Attributes option with Inspect Resource in resource context menu, leveraging the updated CICS Resource Inspector [#422](https://github.com/zowe/cics-for-zowe-client/issues/422)
+- Removal of RI Actions component for unsupported Resources [#423](https://github.com/zowe/cics-for-zowe-client/issues/423)
 
 ## `3.12.0`