+++ packages/vsce/CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 
 ## Recent Changes
 
+- Enhancement: Added JVM Endpoint to the JVM Server node. [#410](https://github.com/zowe/cics-for-zowe-client/issues/413)
 - Removal of RI Actions component for unsupported Resources [#423](https://github.com/zowe/cics-for-zowe-client/issues/423)
 
 ## `3.12.0`
 
-- Enhancement: Added JVM Endpoint to the JVM Server node. [#413](https://github.com/zowe/cics-for-zowe-client/issues/413)
 - Validate input using maximum resource length when inspecting resource. [#402](https://github.com/zowe/cics-for-zowe-client/issues/402)
 - Action menus on Resource Inspector Header. [#333](https://github.com/zowe/cics-for-zowe-client/issues/333)
 - Enhancement: Added Enable/Disable actions to JVM Servers [#391](https://github.com/zowe/cics-for-zowe-client/issues/391)