+++ packages/vsce/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 
 ## Recent Changes
 
+- Replaced the Show Attributes option with Inspect Resource in resource context menu, leveraging the updated CICS Resource Inspector [#422](https://github.com/zowe/cics-for-zowe-client/issues/422)
 - Removal of RI Actions component for unsupported Resources [#423](https://github.com/zowe/cics-for-zowe-client/issues/423)
 - Enhancement: Added JVM Endpoint to the JVM Server node [#413](https://github.com/zowe/cics-for-zowe-client/issues/413)